+++ pages/options.js
@@ -1,3 +1,8 @@
+import "./all_content_scripts.js";
+import { ExclusionRulesEditor } from "./exclusion_rules_editor.js";
+import { allCommands } from "../background_scripts/all_commands.js";
+import { Commands } from "../background_scripts/commands.js";
+
 const options = {
   filterLinkHints: "boolean",
   waitForEnterForFilteredHints: "boolean",
@@ -23,14 +28,14 @@ const OptionsPage = {
   async init() {
     await Settings.onLoaded();
 
-    const saveOptionsEl = document.querySelector("#saveOptions");
+    const saveButton = document.querySelector("#save");
 
     const isMacOS = KeyboardUtils.platform === "Mac";
     if (isMacOS) document.getElementById("macos").style.display = "contents";
 
     const onUpdated = () => {
-      saveOptionsEl.disabled = false;
-      saveOptionsEl.textContent = "Save changes";
+      saveButton.disabled = false;
+      saveButton.textContent = "Save changes";
     };
 
     for (const el of document.querySelectorAll("input, textarea")) {
@@ -42,23 +47,19 @@ const OptionsPage = {
       });
     }
 
-    saveOptionsEl.addEventListener("click", () => this.saveOptions());
-    document.querySelector("#showCommands").addEventListener(
-      "click",
-      () => HelpDialog.toggle({ showAllCommandDetails: true }),
-    );
+    saveButton.addEventListener("click", () => this.saveOptions());
 
-    document.querySelector("#filterLinkHints").addEventListener(
+    this.getOptionEl("filterLinkHints").addEventListener(
       "click",
       () => this.maintainLinkHintsView(),
     );
 
-    document.querySelector("#downloadBackup").addEventListener(
+    document.querySelector("#download-backup").addEventListener(
       "mousedown",
       () => this.onDownloadBackupClicked(),
       true,
     );
-    document.querySelector("#uploadBackup").addEventListener(
+    document.querySelector("#upload-backup").addEventListener(
       "change",
       () => this.onUploadBackupClicked(),
     );
@@ -72,7 +73,7 @@ const OptionsPage = {
     }
 
     globalThis.onbeforeunload = () => {
-      if (!saveOptionsEl.disabled) {
+      if (!saveButton.disabled) {
         return "You have unsaved changes to options.";
       }
     };
@@ -91,12 +92,16 @@ const OptionsPage = {
     this.setFormFromSettings(settings);
   },
 
+  getOptionEl(optionName) {
+    return document.querySelector(`*[name="${optionName}"]`);
+  },
+
   // Invoked when the user clicks the "reset" button next to an option's text field.
   resetInputValue(event) {
     const parentDiv = event.target.parentNode.parentNode;
     console.assert(parentDiv?.tagName == "DIV", "Expected parent to be a div", event.target);
     const input = parentDiv.querySelector("input") || parentDiv.querySelector("textarea");
-    const optionName = input.id;
+    const optionName = input.name;
     const defaultValue = Settings.defaultOptions[optionName];
     input.value = defaultValue;
     event.preventDefault();
@@ -104,7 +109,7 @@ const OptionsPage = {
 
   setFormFromSettings(settings) {
     for (const [optionName, optionType] of Object.entries(options)) {
-      const el = document.getElementById(optionName);
+      const el = this.getOptionEl(optionName);
       const value = settings[optionName];
       switch (optionType) {
         case "boolean":
@@ -123,14 +128,14 @@ const OptionsPage = {
 
     ExclusionRulesEditor.setForm(Settings.get("exclusionRules"));
 
-    document.querySelector("#uploadBackup").value = "";
+    document.querySelector("#upload-backup").value = "";
     this.maintainLinkHintsView();
   },
 
   getSettingsFromForm() {
     const settings = {};
     for (const [optionName, optionType] of Object.entries(options)) {
-      const el = document.getElementById(optionName);
+      const el = this.getOptionEl(optionName);
       let value;
       switch (optionType) {
         case "boolean":
@@ -161,21 +166,21 @@ const OptionsPage = {
     let text, parsed;
 
     // keyMappings field.
-    text = document.getElementById("keyMappings").value.trim();
+    text = this.getOptionEl("keyMappings").value.trim();
     parsed = Commands.parseKeyMappingsConfig(text);
     if (parsed.validationErrors.length > 0) {
       results["keyMappings"] = parsed.validationErrors.join("\n");
     }
 
     // searchEngines field.
-    text = document.getElementById("searchEngines").value.trim();
+    text = this.getOptionEl("searchEngines").value.trim();
     parsed = UserSearchEngines.parseConfig(text);
     if (parsed.validationErrors.length > 0) {
       results["searchEngines"] = parsed.validationErrors.join("\n");
     }
 
     // linkHintCharacters field.
-    text = document.getElementById("linkHintCharacters").value.trim();
+    text = this.getOptionEl("linkHintCharacters").value.trim();
     if (text != this.removeDuplicateChars(text)) {
       results["linkHintCharacters"] = "This cannot contain duplicate characters.";
     } else if (text.length <= 1) {
@@ -183,7 +188,7 @@ const OptionsPage = {
     }
 
     // linkHintNumbers field.
-    text = document.getElementById("linkHintNumbers").value.trim();
+    text = this.getOptionEl("linkHintNumbers").value.trim();
     if (text != this.removeDuplicateChars(text)) {
       results["linkHintNumbers"] = "This cannot contain duplicate characters.";
     } else if (text.length <= 1) {
@@ -198,7 +203,7 @@ const OptionsPage = {
     const exampleEl = el.nextElementSibling;
     const messageEl = document.createElement("div");
     messageEl.classList.add("validation-message");
-    messageEl.innerText = message;
+    messageEl.textContent = message;
     exampleEl.after(messageEl);
   },
 
@@ -216,15 +221,15 @@ const OptionsPage = {
 
     const errors = this.getValidationErrors();
     for (const [optionName, message] of Object.entries(errors)) {
-      const el = document.getElementById(optionName);
+      const el = this.getOptionEl(optionName);
       this.addValidationMessage(el, message);
     }
     // Some options can be hidden in the UI. If they have validation errors, force them to be shown.
     if (errors["linkHintCharacters"]) {
-      this.showElement(document.querySelector("#linkHintCharactersContainer"), true);
+      this.showElement(document.querySelector("#link-hint-characters-container"), true);
     }
     if (errors["linkHintNumbers"]) {
-      this.showElement(document.querySelector("#linkHintNumbersContainer"), true);
+      this.showElement(document.querySelector("#link-hint-numbers-container"), true);
     }
     const hasErrors = Object.keys(errors).length > 0;
     return hasErrors;
@@ -251,7 +256,7 @@ const OptionsPage = {
     }
 
     await Settings.setSettings(this.getSettingsFromForm());
-    const el = document.querySelector("#saveOptions");
+    const el = document.querySelector("#save");
     el.disabled = true;
     el.textContent = "Saved";
   },
@@ -264,17 +269,17 @@ const OptionsPage = {
   // "filterLinkHints".
   maintainLinkHintsView() {
     const errors = this.getValidationErrors();
-    const isFilteredLinkhints = document.querySelector("#filterLinkHints").checked;
+    const isFilteredLinkhints = this.getOptionEl("filterLinkHints").checked;
     this.showElement(
-      document.querySelector("#linkHintCharactersContainer"),
+      document.querySelector("#link-hint-characters-container"),
       !isFilteredLinkhints || errors["linkHintCharacters"],
     );
     this.showElement(
-      document.querySelector("#linkHintNumbersContainer"),
+      document.querySelector("#link-hint-numbers-container"),
       isFilteredLinkhints || errors["linkHintNumbers"],
     );
     this.showElement(
-      document.querySelector("#waitForEnterForFilteredHintsContainer"),
+      document.querySelector("#wait-for-enter"),
       isFilteredLinkhints,
     );
   },
@@ -282,7 +287,7 @@ const OptionsPage = {
   onDownloadBackupClicked() {
     const backup = Settings.pruneOutDefaultValues(this.getSettingsFromForm());
     const settingsBlob = new Blob([JSON.stringify(backup, null, 2) + "\n"]);
-    document.querySelector("#downloadBackup").href = URL.createObjectURL(settingsBlob);
+    document.querySelector("#download-backup").href = URL.createObjectURL(settingsBlob);
   },
 
   onUploadBackupClicked() {
@@ -307,9 +312,9 @@ const OptionsPage = {
 
         await Settings.setSettings(backup);
         this.setFormFromSettings(Settings.getSettings());
-        const saveOptionsEl = document.querySelector("#saveOptions");
-        saveOptionsEl.disabled = true;
-        saveOptionsEl.textContent = "Saved";
+        const saveButton = document.querySelector("#save");
+        saveButton.disabled = true;
+        saveButton.textContent = "Saved";
         alert("Settings have been restored from the backup.");
       };
     }