+++ tests/unit_tests/commands_test.js
@@ -1,9 +1,8 @@
 import "./test_helper.js";
-import "../../background_scripts/tab_recency.js";
-import "../../background_scripts/bg_utils.js";
 import "../../lib/settings.js";
 import "../../lib/keyboard_utils.js";
-import "../../background_scripts/commands.js";
+import { allCommands } from "../../background_scripts/all_commands.js";
+import { Commands, defaultKeyMappings } from "../../background_scripts/commands.js";
 import "../../content_scripts/mode.js";
 import "../../content_scripts/mode_key_handler.js";
 // Include mode_normal to check that all commands have been implemented.
@@ -131,40 +130,33 @@ context("parseKeyMappingConfig", () => {
   });
 });
 
-context("Validate commands and options", () => {
-  // TODO(smblott) For this and each following test, is there a way to structure the tests such that the name
-  // of the offending command appears in the output, if the test fails?
+context("Validate commands and options data structures", () => {
   should("have either noRepeat or repeatLimit, but not both", () => {
-    for (const command of Object.keys(Commands.availableCommands)) {
-      const options = Commands.availableCommands[command];
-      assert.isTrue(!(options.noRepeat && options.repeatLimit));
+    for (const command of allCommands) {
+      const validProperties = !(command.noRepeat && command.repeatLimit);
+      if (!validProperties) {
+        assert.fail(`${command.name} has incorrect noRepeat and/or repeatLimit config.`);
+      }
     }
   });
 
-  should("describe each command", () => {
-    for (const command of Object.keys(Commands.availableCommands)) {
-      const options = Commands.availableCommands[command];
-      assert.equal("string", typeof options.description);
-    }
-  });
-
-  should("define each command in each command group", () => {
-    for (const group of Object.keys(Commands.commandGroups)) {
-      const commands = Commands.commandGroups[group];
-      for (const command of commands) {
-        assert.equal("string", typeof command);
-        assert.isTrue(Commands.availableCommands[command]);
+  should("have required properties", () => {
+    for (const command of allCommands) {
+      const hasRequired = command.desc.length > 0 && command.group.length > 0;
+      if (!hasRequired) {
+        assert.fail(`${command.name} is missing required properties.`);
       }
     }
   });
 
   should("have valid commands for each default key mapping", () => {
-    const count = Object.keys(Commands.keyToRegistryEntry).length;
-    assert.isTrue(0 < count);
-    for (const key of Object.keys(Commands.keyToRegistryEntry)) {
-      const command = Commands.keyToRegistryEntry[key];
-      assert.equal("object", typeof command);
-      assert.isTrue(Commands.availableCommands[command.command]);
+    const commandsByName = Utils.keyBy(allCommands, "name");
+    for (const [key, commandString] of Object.entries(defaultKeyMappings)) {
+      // The comamnd string might be command name + an option string. Ignore the options.
+      const name = commandString.split(" ")[0];
+      if (commandsByName[name] == null) {
+        assert.fail(`The default mapping for ${key} is bound to non-existant command ${name}.`);
+      }
     }
   });
 });