+++ tests/unit_tests/marks_test.js
@@ -1,12 +1,12 @@
 import "./test_helper.js";
-import "../../background_scripts/marks.js";
+import * as marks from "../../background_scripts/marks.js";
 
 context("marks", () => {
   const createMark = async (markProperties, tabProperties) => {
     const mark = Object.assign({ scrollX: 0, scrollY: 0 }, markProperties);
     const tab = Object.assign({ url: "http://example.com" }, tabProperties);
     const sender = { tab: tab };
-    await Marks.create(mark, sender);
+    await marks.create(mark, sender);
   };
 
   setup(() => {
@@ -21,7 +21,7 @@ context("marks", () => {
 
   should("record the vimium secret in the mark's info", async () => {
     await createMark({ markName: "a" });
-    const key = Marks.getLocationKey("a");
+    const key = marks.getLocationKey("a");
     const savedMark = (await chrome.storage.local.get(key))[key];
     assert.equal("secret", savedMark.vimiumSecret);
   });
@@ -32,7 +32,7 @@ context("marks", () => {
     stub(globalThis.chrome.tabs, "get", (id) => id == 1 ? tab : null);
     const updatedTabs = [];
     stub(globalThis.chrome.tabs, "update", (id, properties) => updatedTabs[id] = properties);
-    await Marks.goto({ markName: "A" });
+    await marks.goto({ markName: "A" });
     assert.isTrue(updatedTabs[1] && updatedTabs[1].active);
   });
 
@@ -45,7 +45,7 @@ context("marks", () => {
     stub(globalThis.chrome.tabs, "query", (_) => [tab]);
     const updatedTabs = [];
     stub(globalThis.chrome.tabs, "update", (id, properties) => updatedTabs[id] = properties);
-    await Marks.goto({ markName: "A" });
+    await marks.goto({ markName: "A" });
     assert.isTrue(updatedTabs[2] && updatedTabs[2].active);
   });
 });