+++ tests/unit_tests/vomnibar_test.js
@@ -1,9 +1,9 @@
-import { assert, context, setup, should, stub, teardown } from "../vendor/shoulda.js";
-import * as shoulda from "../vendor/shoulda.js";
-import * as jsdom from "npm:jsdom";
+import * as testHelper from "./test_helper.js";
 import "../../tests/unit_tests/test_chrome_stubs.js";
+
+import { Suggestion } from "../../background_scripts/completion.js";
 import "../../background_scripts/completion.js";
-import { Vomnibar } from "../../pages/vomnibar.js";
+import { Vomnibar } from "../../pages/vomnibar_page.js";
 
 function newKeyEvent(properties) {
   return Object.assign(
@@ -23,20 +23,12 @@ function newKeyEvent(properties) {
 
 context("vomnibar", () => {
   setup(async () => {
-    const html = await Deno.readTextFile("pages/vomnibar.html");
-    const w = new jsdom.JSDOM(html).window;
-    globalThis.window = w;
-    globalThis.document = w.document;
+    await testHelper.jsdomStub("pages/vomnibar_page.html");
     stub(chrome.runtime, "sendMessage", async (message) => {
       if (message.handler == "filterCompletions") {
         return [];
       }
     });
-  });
-
-  teardown(() => {
-    globalThis.window = undefined;
-    globalThis.document = undefined;
   });
 
   should("hide when escape is pressed", async () => {