+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+2.3 (unreleased)
+
+- Some internal CSS classes were changed for Vimium's UI. This may affect those who have customized
+  Vimium's CSS via the options page. [#4668](https://github.com/philc/vimium/issues/4668).
+
 2.2.1 (2025-03-20)
 
 - Fix findSelected and findSelectedBackwards commands (https://github.com/philc/vimium/issues/4655)